add debug super-endpoint (#33)
We wish to re-use the functionality from `ProfileEndpoints` to upload the profiles to blob storage if our server is running on a specific environment.

Let's add a `/debug_engine` super-endpoint which is responsible for profiling an engine and writing the results to a directory specified at server initialization (`stage_path`).

The type of profile (e.g. CPU profile, heap snapshot, etc.) is specified in the `profile_type` parameter, which is used to de-multiplex the request which hits the `debug_engine` endpoint to the appropriate profile function.